+++ packages/tween/src/index.ts
@@ -2,3 +2,6 @@ import * as easings from './easing';
 
 export { easings };
 export { tween } from './helper';
+export { Tween } from './tween';
+export { Timeline } from './timeline';
+export { TweenController } from './controller';